--- src/shared/newGame/types.ts
@@ -1,24 +1,0 @@
-export enum ScoringMode {
-  Standard,
-  Negative,
-}
-
-export interface IGameSettings {
-  scoringMode: ScoringMode;
-  bonusRounds: boolean;
-}
-
-export interface INewGameRequest {
-  playerNames: string[];
-  dealer: string;
-  settings: IGameSettings;
-}
-
-export const defaultRequest: INewGameRequest = {
-  playerNames: [],
-  dealer: '',
-  settings: {
-    bonusRounds: true,
-    scoringMode: ScoringMode.Negative,
-  },
-};
