+++ src/shared/persistence/types.ts
@@ -1,4 +1,27 @@
-import { INewGameRequest, IGameSettings, defaultRequest } from '../newGame/types';
+export enum ScoringMode {
+  Standard,
+  Negative,
+}
+
+export interface IGameSettings {
+  scoringMode: ScoringMode;
+  bonusRounds: boolean;
+}
+
+export interface INewGameRequest {
+  playerNames: string[];
+  dealer: string;
+  settings: IGameSettings;
+}
+
+export const defaultRequest: INewGameRequest = {
+  playerNames: [],
+  dealer: '',
+  settings: {
+    bonusRounds: true,
+    scoringMode: ScoringMode.Negative,
+  },
+};
 
 export interface IGame extends INewGameRequest {
   dealer: string;
